support teammate cursors
