+++ src/shared/gameEvents/types/GameState.ts
@@ -1,14 +1,9 @@
-import {CellCoords, GameJson} from '../../types';
+import {GameJson, UserJson} from '../../types';
 
 export interface GameState {
   loaded: boolean;
   game: null | GameJson;
   users: {
-    [id: string]: {
-      cursor?: CellCoords;
-      displayName: string;
-      teamId?: number;
-      score?: number;
-    };
+    [id: string]: UserJson;
   };
 }
