+++ src/shared/types.ts
@@ -33,6 +33,23 @@ export interface GameJson {
   solution: string[][];
   clues: CluesJson;
 }
+
+export interface UserJson {
+  cursor?: Cursor;
+  displayName: string;
+  teamId?: number;
+  score?: number;
+}
+
+export interface Cursor {
+  id: string;
+  r: number; // Row in puzzle
+  c: number; // Column in puzzle
+  timestamp: number;
+  color?: string;
+  active?: boolean;
+}
+
 /**
  * PuzzleJson: the json format of puzzles stored in the db (both firebase & postgres)
  * Fields are a bit messy & don't correspond perfectly with puzjs formats... see logic in FileUploader.js
